pyproject.toml
[project]
name = "obsidian-vault-server"
version = "0.1.0"
description = "MCP/REST server for Obsidian vault management with semantic search and knowledge graph"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Web Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.18",

    # MCP Server
    "fastmcp>=0.4.0",

    # Database
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pgvector>=0.3.6",

    # Authentication
    "pyjwt>=2.10.0",
    "argon2-cffi>=23.1.0",

    # Configuration
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",

    # OpenAI Embeddings
    "openai>=1.57.0",
    "tiktoken>=0.8.0",

    # Utilities
    "pyyaml>=6.0.2",
    "python-slugify>=8.0.4",
    "httpx>=0.28.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-bdd>=8.1.0",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "httpx>=0.28.0",
    "factory-boy>=3.3.0",

    # Linting & Formatting
    "ruff>=0.8.0",
    "mypy>=1.13.0",

    # Type stubs
    "types-pyyaml>=6.0.12",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "bdd: BDD tests",
]

[tool.coverage.run]
source = ["app"]
branch = true
omit = ["app/infrastructure/database/migrations/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
fail_under = 70

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["pgvector.*", "asyncpg.*"]
ignore_missing_imports = true
